refactor(search): move serde_fields inline tests to external test file
serde_fields.rs exceeded the 400-line project guideline at 477 lines,
with 170 lines being inline unit tests. Extracted all #[cfg(test)] tests
to tests/serde_field_classification.rs and renamed them with a low_level_
prefix to avoid collisions with existing integration tests.

Source file is now 302 lines. All 43 tests pass.
